+++ src/index.html
@@ -75,7 +75,7 @@
 
     </div>
   
-  <script src="main2.js" type="module"></script>
+  <script src="main.js" type="module"></script>
   <!-- <script src="data.js" type="module"></script> -->
   </body>
   </html>
