+++ test/data.spec.js
@@ -1,5 +1,39 @@
-import {dataFiltradoRol,dataFiltradoTop,dataFiltradoOrden} from '../src/metodos2.js';
+import {dataFiltradoRol,dataFiltradoTop,dataFiltradoOrden,mostrarConteo} from '../src/metodos.js';
   
+describe('dataFiltradoRol', () => {
+  it('should return []', () => {
+    expect(dataFiltradoRol('bard').length).toBe(0);
+  });
+
+  it('should return 50', () => {
+    expect(dataFiltradoRol('Mage').length).toBe(50);
+  });
+
+  it('should return 24', () => {
+    expect(dataFiltradoRol('Marksman').length).toBe(24);
+  });
+  
+})
+
+describe('dataFiltradoTop', () => {
+  // it('should return []', () => {
+  //   expect(dataFiltradoTop('PA').length).toBe(0);
+  // });
+
+  // it('should return 10', () => {
+  //   expect(dataFiltradoTop('HP').length).toBe(10);
+  // });
+
+  it('should return 10', () => {
+    expect(dataFiltradoTop('RA').length).toBe(10);
+  });
+  
+})
+
+
+
+
+
 
 describe('Testear Objeto Funcion', () => {
   it('funcion dataFiltradoRol', () => {
@@ -14,7 +48,12 @@ describe('Testear Objeto Funcion', () => {
     expect(typeof dataFiltradoOrden).toBe('function');
   });
 
+  it('funcion dataFiltradoOrden', () => {
+    expect(typeof mostrarConteo).toBe('function');
+  });
+
 });
+
 
 
 // describe('anotherExample', () => {
